chore: extract derivedStateFromProps
